feat: Add poisoned_nodes attribute to data objects
The `poisoned_nodes` attribute is added to the `data` objects in order to keep track of the indices of the poisoned nodes. This attribute is used in various parts of the codebase to perform operations related to poisoning and unlearning. This change ensures that the `poisoned_nodes` attribute is properly set and utilized throughout the codebase.